Update patch set 1
Patch Set 1:

If you try building it, you'll see we'd have to forward port it to Oreo libraries changes, and Google killed the sources on purpose to break old packages that might reference libstlport, and to avoid maintaining dead code that won't be used in the future anyways.

Only legacy vendor libs depend on stlport and the prebuilt that goes along, for instance flo for ARM32, matches and works just fine for that.

Patch-set: 1